meiti/requirements.txt
# Core
python-telegram-bot[job-queue]==20.6
python-dotenv==1.0.0

# Web
Flask==3.0.0
aiohttp==3.9.1
aiosqlite==0.19.0

# Utils
qrcode==7.4.2
Pillow==10.1.0
